+++ src/main/java/io/coodoo/framework/listing/boundary/Listing.java
@@ -4,26 +4,29 @@ import java.util.List;
 
 import javax.persistence.EntityManager;
 
-import io.coodoo.framework.listing.control.ListingFilterQuery;
+import io.coodoo.framework.listing.control.ListingQuery;
 
+/**
+ * @author coodoo GmbH (coodoo.io)
+ */
 public class Listing {
 
-    public static <T> ListingResult<T> getListingResult(EntityManager entityManager, Class<T> entityClass, ListingQueryParams queryParams) {
+    public static <T> ListingResult<T> getListingResult(EntityManager entityManager, Class<T> entityClass, ListingParameters queryParams) {
         return new ListingResult<T>(getListing(entityManager, entityClass, queryParams),
                         new Metadata(countListing(entityManager, entityClass, queryParams), queryParams));
     }
 
-    public static <T> List<T> getListing(EntityManager entityManager, Class<T> entityClass, ListingQueryParams queryParams) {
+    public static <T> List<T> getListing(EntityManager entityManager, Class<T> entityClass, ListingParameters queryParams) {
         return assambleFilter(entityManager, entityClass, queryParams).list(queryParams.getIndex(), queryParams.getLimit());
     }
 
-    public static <T> Long countListing(EntityManager entityManager, Class<T> entityClass, ListingQueryParams queryParams) {
+    public static <T> Long countListing(EntityManager entityManager, Class<T> entityClass, ListingParameters queryParams) {
         return assambleFilter(entityManager, entityClass, queryParams).count();
     }
 
-    private static <T> ListingFilterQuery<T> assambleFilter(EntityManager entityManager, Class<T> entityClass, ListingQueryParams queryParams) {
+    private static <T> ListingQuery<T> assambleFilter(EntityManager entityManager, Class<T> entityClass, ListingParameters queryParams) {
 
-        return new ListingFilterQuery<>(entityManager, entityClass)
+        return new ListingQuery<>(entityManager, entityClass)
                         // apply sorting
                         .sort(queryParams.getSortAttribute(), queryParams.isSortAsc())
                         // disjunctive filter on the whole table
@@ -35,27 +38,27 @@ public class Listing {
     }
 
     public static <T> ListingResult<T> getListingResult(EntityManager entityManager, Class<T> entityClass, Integer page, Integer limit) {
-        return getListingResult(entityManager, entityClass, new ListingQueryParams(page, limit, null));
+        return getListingResult(entityManager, entityClass, new ListingParameters(page, limit, null));
     }
 
     public static <T> List<T> getListing(EntityManager entityManager, Class<T> entityClass, Integer page, Integer limit) {
-        return getListing(entityManager, entityClass, new ListingQueryParams(page, limit, null));
+        return getListing(entityManager, entityClass, new ListingParameters(page, limit, null));
     }
 
     public static <T> Long countListing(EntityManager entityManager, Class<T> entityClass, Integer page, Integer limit) {
-        return countListing(entityManager, entityClass, new ListingQueryParams(page, limit, null));
+        return countListing(entityManager, entityClass, new ListingParameters(page, limit, null));
     }
 
     public static <T> ListingResult<T> getListingResult(EntityManager entityManager, Class<T> entityClass, Integer page, Integer limit, String sortAttribute) {
-        return getListingResult(entityManager, entityClass, new ListingQueryParams(page, limit, sortAttribute));
+        return getListingResult(entityManager, entityClass, new ListingParameters(page, limit, sortAttribute));
     }
 
     public static <T> List<T> getListing(EntityManager entityManager, Class<T> entityClass, Integer page, Integer limit, String sortAttribute) {
-        return getListing(entityManager, entityClass, new ListingQueryParams(page, limit, sortAttribute));
+        return getListing(entityManager, entityClass, new ListingParameters(page, limit, sortAttribute));
     }
 
     public static <T> Long countListing(EntityManager entityManager, Class<T> entityClass, Integer page, Integer limit, String sortAttribute) {
-        return countListing(entityManager, entityClass, new ListingQueryParams(page, limit, sortAttribute));
+        return countListing(entityManager, entityClass, new ListingParameters(page, limit, sortAttribute));
     }
 
 }