+++ src/main/java/io/coodoo/framework/listing/boundary/annotation/ListingEntityManager.java
@@ -7,6 +7,9 @@ import java.lang.annotation.Target;
 
 import javax.inject.Qualifier;
 
+/**
+ * @author coodoo GmbH (coodoo.io)
+ */
 @Qualifier
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ElementType.METHOD, ElementType.FIELD})